EducationOrganizationId BIGINT updated

+++ Extensions/EdFi.Ods.Extensions.Sample/Versions/1.0.0/Standard/5.2.0/Artifacts/MsSql/Structure/Ods/0020-EXTENSION-Sample-Tables.sql
@@ -468,7 +468,7 @@ GO
 
 -- Table [sample].[StudentAcademicRecordExtension] --
 CREATE TABLE [sample].[StudentAcademicRecordExtension] (
-    [EducationOrganizationId] [INT] NOT NULL,
+    [EducationOrganizationId] [BIGINT] NOT NULL,
     [SchoolYear] [SMALLINT] NOT NULL,
     [StudentUSI] [INT] NOT NULL,
     [TermDescriptorId] [INT] NOT NULL,
