+++ Extensions/EdFi.Ods.Extensions.Sample/Versions/1.0.0/Standard/5.2.0/Artifacts/PgSql/Structure/Ods/0020-EXTENSION-Sample-Tables.sql
@@ -306,7 +306,7 @@ ALTER TABLE sample.StaffPet ALTER COLUMN CreateDate SET DEFAULT current_timestam
 
 -- Table sample.StudentAcademicRecordExtension --
 CREATE TABLE sample.StudentAcademicRecordExtension (
-    EducationOrganizationId INT NOT NULL,
+    EducationOrganizationId BIGINT NOT NULL,
     SchoolYear SMALLINT NOT NULL,
     StudentUSI INT NOT NULL,
     TermDescriptorId INT NOT NULL,
